feat!: bump to vitepress beta 2
