+++ package.json
@@ -57,7 +57,7 @@
     "unbuild": "^1.2.1",
     "vite": "^4.3.9",
     "vite-plugin-pwa": ">=0.16.3 <1",
-    "vitepress": "1.0.0-beta.1"
+    "vitepress": "1.0.0-beta.2"
   },
   "pnpm": {
     "peerDependencyRules": {
